Simplify steps to load example data (#83)
Resolves https://github.com/dbt-labs/jaffle-shop/issues/82

### Problem

Currently, to load example data, the instructions are to edit the
`dbt_project.yml` once to enable the `jaffle-data` seeds and then
another time to disable them. This is a bit cumbersome for my taste!

The repo is explicitly constructed in a way to make it easy to delete
all the `jaffle-data` CSVs because we don't want to encourage using dbt
as a data loading tool. These seed files only exist to load example
data.

### Solution

Use `--vars` CLI + `var` macro + `enabled` config to opt-in to loading
CSV seeds, but by default do not include them in `dbt build`.